- Bugzilla Bug (s) Resolved:
- Description: 
In previous change, the combo list was used for some chart properties
and added an 'auto' item in combo list, but it is not friendly. Now we
use tristate checkbox instead, the Grayed state of tristate checkbox
indicates 'auto' case.(#39912)

+++ chart/org.eclipse.birt.chart.ui.extension/src/org/eclipse/birt/chart/ui/swt/series/PieSeriesAttributeComposite.java
@@ -500,7 +500,7 @@ public class PieSeriesAttributeComposite extends Composite implements
 		{
 			ChartElementUtil.setEObjectAttribute( series,
 					"leaderLineLength", //$NON-NLS-1$
-					iscLeaderLength.getSelection( ),
+					(double)iscLeaderLength.getSelection( ),
 					btnLeaderLengthAuto.getSelection( ) );
 			iscLeaderLength.setEnabled( !btnLeaderLengthAuto.getSelection( ) );
 		}
